+++ threedigrid_builder/application.py
@@ -9,10 +9,11 @@ from pathlib import Path
 from threedigrid_builder.exceptions import SchematisationError
 from threedigrid_builder.grid import Grid
 from threedigrid_builder.grid import QuadTree
+from threedigrid_builder.interface import GDALInterface
 from threedigrid_builder.interface import GeopackageOut
 from threedigrid_builder.interface import GridAdminOut
+from threedigrid_builder.interface import RasterioInterface
 from threedigrid_builder.interface import SQLite
-from threedigrid_builder.interface import Subgrid
 from typing import Callable
 from typing import Optional
 
@@ -32,7 +33,6 @@ def _default_progress_callback(progress: float, message: str):
 def _make_gridadmin(
     sqlite_path,
     dem_path=None,
-    model_area_path=None,
     meta=None,
     progress_callback=None,
 ):
@@ -53,8 +53,17 @@ def _make_gridadmin(
         if not dem_path:
             raise SchematisationError("DEM file expected")
         # TODO use_2d_flow --> https://github.com/nens/threedigrid-builder/issues/87
-        subgrid = Subgrid(dem_path, model_area=model_area_path)
-        subgrid_meta = subgrid.get_meta()
+
+        try:
+            with RasterioInterface(dem_path) as raster:
+                subgrid_meta = raster.read()
+        except ImportError:
+            with GDALInterface(dem_path) as raster:
+                subgrid_meta = raster.read()
+
+        # Patch epsg code with that of the DEM (so: user-supplied EPSG is ignored)
+        grid.meta.epsg_code = raster.epsg_code
+
         refinements = db.get_grid_refinements()
         progress_callback(0.7, "Constructing 2D computational grid...")
         quadtree = QuadTree(
@@ -211,7 +220,6 @@ def make_gridadmin(
     sqlite_path: Path,
     dem_path: Path,
     out_path: Path,
-    model_area_path: Optional[Path] = None,
     meta: dict = None,
     progress_callback: Optional[Callable[[float, str], None]] = None,
 ):
@@ -224,7 +232,6 @@ def make_gridadmin(
         dem_path: The path of the input DEM file (GeoTIFF)
         out_path: The path of the (to be created) output file. Allowed extensions
             are: .h5 (HDF5) and .gpkg (Geopackage)
-        model_area_path
         meta: an optional dict with the following (optional) keys: model_slug (str),
             revision_hash (str), revision_nr (int), threedi_version (str)
         progress_callback: an optional function that updates the progress. The function
@@ -240,8 +247,6 @@ def make_gridadmin(
         dem_path = Path(dem_path)
     if isinstance(out_path, str):
         out_path = Path(out_path)
-    if isinstance(model_area_path, str):
-        model_area_path = Path(model_area_path)
     extension = out_path.suffix.lower()
     if extension == ".h5":
         writer = _grid_to_hdf5
@@ -255,7 +260,6 @@ def make_gridadmin(
     grid = _make_gridadmin(
         sqlite_path,
         dem_path,
-        model_area_path,
         meta=meta,
         progress_callback=progress_callback,
     )