+++ threedigrid_builder/interface/__init__.py
@@ -10,6 +10,7 @@ The interface layer depends on the domain layer only.
 """
 
 from .db import *  # NOQA
-from .subgrid import *  # NOQA
 from .geopackage import *  # NOQA
 from .gridadmin import *  # NOQA
+from .raster_rasterio import *  # NOQA
+from .raster_gdal import *  # NOQA