+++ CHANGES.rst
@@ -8,6 +8,12 @@ Changelog of threedigrid-builder
 
 - Added breaches and levees.
 
+- Implement GDAL as an alternative to RasterIO.
+
+- Check the raster EPSG code and use it if the model does not have one.
+
+- Removed 'model_area_path' feature from application.
+
 
 0.7.0 (2021-11-25)
 ------------------